requirements.txt
geopandas
sqlalchemy
boto3
pyarrow
pandas
pyyaml
psycopg2-binary